% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{kappaRule_metrics}
\alias{kappaRule_metrics}
\title{Function to evaluate the model}
\usage{
kappaRule_metrics(test_data, model, measure = NULL)
}
\arguments{
\item{test_data}{test data to evaluate the model on}

\item{model}{saved kappaRule model}
}
\description{
Function to evaluate the model
}
