% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{predict.ruleFit}
\alias{predict.ruleFit}
\title{Predict function for varppRule}
\usage{
\method{predict}{ruleFit}(test_data, model, predict = c("probability", "class"))
}
\arguments{
\item{test_data}{a test data set}

\item{model}{the stored kappaRule model}

\item{predict}{if the prediction should be probability or class}
}
\description{
Predict function for varppRule
}
